GDP, Population, and GDP per Capita (Correlates of War System)

These are estimates of democracy for Correlates of War state system members.

Usage

cw_gdppop

Format

A data frame with the following 8 variables.

ccode

a numeric vector for the Correlates of War state code

year

a numeric vector for the year

mrgdppc

a numeric vector for the estimated GDP per capita in a given year. See Details.

sd_mrgdppc

a numeric vector for the standard deviation of estimated GDP per capita in a given year.

pwtrgdp

a numeric vector for the estimated real GDP in a given year. See Details.

sd_pwtrgdp

a numeric vector for the standard deviation of estimated real GDP in a given year.

pwtpop

a numeric vector for the estimated population in a given year. See Details.

sd_pwtpop

a numeric vector for the standard deviation of estimated population in a given year.

Details

Fariss et al. (2022) use Gleditsch-Ward for their population of cases. The differences between Gleditsch-Ward and Correlates of War are obvious if often overstated. However, there will be cases where merging one system into the other amounts to a collision, the wreckage of which can't go unnoticed. The canonical cases here are post-WW2 Germany, post-unification Yemen, and all of Serbia/Yugoslavia. Those merit further scrutiny by the user.

The underlying data, as they are, at at the mercy of the Gleditsch-Ward system for describing the universe of cases that could have a GDP, a population, or a GDP per capita. That means there are missing data for Serbia (1916, 1917), Morocco (1905-1912), Egypt (1856-1882), Saudi Arabia (1927-1931), and Laos (1953). I can think of a few imputation procedures under those circumstances, but that is something for which the user would have to take initiative to do themselves.

Based on my reading of Fariss et al. (2022), I think the following information gathered from their simulations make sense for suggested defaults. You may want to get their actual simulations if you want something else, but I think what's included here is good for most use cases.

For additional clarification, the suggested defaults included in this data set are:

GDP per capita: real GDP per capita in prices constant across countries and over time (in 2011 international dollars, PPP).

GDP: expenditure-side real GDP in prices constant across countries and over time (in millions of 2017 international dollars, PPP)

Population: total population (in millions)

The GDP per capita measure is anchored around the Maddison Project Database. The GDP and population measures are anchored around Penn World Tables (10.0). You can create a rough estimate of GDP per capita from the Penn World Table simulations based on the information in this data set. It's free and the cops can't stop you.

I also honor the authors' please to include the standard deviation of these estimates as well.

References

Please cite Miller (2022) for peacesciencer. Cite Fariss et al. (2022) for the simulations. You should also cite the Maddison Project Database (Bolt et al. 2018) and Penn World Table (Feenstra et al. 2015) if that is the underlying source of the data that Fariss et al. (2022) are estimating.

Bolt, Jutta, Robert Inklaar, Herman de Jong, and Luiten Janvan Zanden. 2018. "Rebasing 'Maddison': New Income Comparisons and the Shape of Long-Run Economic Development." Maddison Project Working paper 10.

Fariss, Christopher, J., Therese Anders, Jonathan N. Markowitz, and Miriam Barnum. 2022. "New Estimates of Over 500 Years of Historic GDP and Population Data." Journal of Conflict Resolution 66(3): 553–91.

Feenstra, Robert C., Robert Inklaar, and Marcel P. Timmer. 2015. "The Next Generation of the Penn World Table." American Economic Review 105(10): 3150–82.