Create a Panel of State-Years from the Correlates of War or Gleditsch-Ward system.

state_panel() is a function to create a panel of state-years from one of two major state systems in international relations scholarship.

Usage

state_panel(system = "cow")

Arguments

system

a state system (either "cow" or "gw")

Value

state_panel() returns a data frame of state years corresponding with either the Correlates of War or the Gleditsch-Ward system.

Details

This function leans on cw_system and gw_system in this package.

Examples

head(state_panel(), 10)
#>    ccode                  cw_name year
#> 1      2 United States of America 1816
#> 2      2 United States of America 1817
#> 3      2 United States of America 1818
#> 4      2 United States of America 1819
#> 5      2 United States of America 1820
#> 6      2 United States of America 1821
#> 7      2 United States of America 1822
#> 8      2 United States of America 1823
#> 9      2 United States of America 1824
#> 10     2 United States of America 1825
head(state_panel(system='gw'), 10)
#>    gwcode                  gw_name year
#> 1       2 United States of America 1816
#> 2       2 United States of America 1817
#> 3       2 United States of America 1818
#> 4       2 United States of America 1819
#> 5       2 United States of America 1820
#> 6       2 United States of America 1821
#> 7       2 United States of America 1822
#> 8       2 United States of America 1823
#> 9       2 United States of America 1824
#> 10      2 United States of America 1825